+++ components/style/lib.rs
@@ -43,7 +43,6 @@ extern crate atomic_refcell;
 extern crate bitflags;
 #[cfg(feature = "gecko")] #[macro_use] #[no_link] extern crate cfg_if;
 #[macro_use] extern crate cssparser;
-extern crate encoding;
 extern crate euclid;
 extern crate fnv;
 #[cfg(feature = "gecko")] #[macro_use] pub mod gecko_string_cache;
@@ -89,6 +88,7 @@ pub mod custom_properties;
 pub mod data;
 pub mod dom;
 pub mod element_state;
+#[cfg(feature = "servo")] mod encoding_support;
 pub mod error_reporting;
 pub mod font_face;
 pub mod font_metrics;
